double to float + handled beyond zeros condition

+++ mini-project/without malloc/evaluate.h
@@ -19,7 +19,7 @@
 
 // Stack definition
 typedef struct {
-    double items[MAX_STACK_SIZE];
+    float items[MAX_STACK_SIZE];
     int top;
 } Stack;
 
@@ -30,28 +30,28 @@ typedef struct {
 extern char __EXPR__[EXP_LEN];                  // expression to be evaluated (x)       - USER SHOULD CHANGE THIS
 extern char __EXPR_VAL__[EXP_LEN];              // expression to be evaluated (number)  - USER SHOULD NOT USE THIS
 extern int __BRACKET_FLAG__;                    // flag to check if brackets are added  - USER SHOULD NOT USE THIS
-extern double __X_MIN__;                        // minimum x value                      - USER SHOULD CHANGE THIS
-extern double __X_MAX__;                        // maximum x value                      - USER SHOULD CHANGE THIS
-extern double __Y_MIN__;                        // minimum y value                      - USER SHOULD USE THIS
-extern double __Y_MAX__;                        // maximum y value                      - USER SHOULD USE THIS
-extern double __Y_MIN_DY_DX__;                  // minimum y value of dy/dx             - USER SHOULD USE THIS
-extern double __Y_MAX_DY_DX__;                  // maximum y value of dy/dx             - USER SHOULD USE THIS
-extern double __Y_MIN_INTEGRAL__;               // minimum y value of integral of y     - USER SHOULD USE THIS
-extern double __Y_MAX_INTEGRAL__;               // maximum y value of integral of y     - USER SHOULD USE THIS
-extern double __MP_X1__;                        // x1 value for mapping                 - USER SHOULD CHANGE THIS
-extern double __MP_Y1__;                        // y1 value for mapping                 - USER SHOULD CHANGE THIS
-extern double __MP_X2__;                        // x2 value for mapping                 - USER SHOULD CHANGE THIS
-extern double __MP_Y2__;                        // y2 value for mapping                 - USER SHOULD CHANGE THIS
-extern double __ZOOMING__;                      // zooming factor                       - USER SHOULD NOT CHANGE THIS
-extern double __SHIFTING__;                     // shifting factor                      - USER SHOULD NOT CHANGE THiS
-extern double __XY__[2][N];                     // x values                             - USER SHOULD CAN USE THIS
-extern double __MAPPED_XY__[2][N];              // x values mapped to screen coordinates- USER SHOULD USE THIS
-extern double __DY_DX__[2][N-1];                // derivative values                    - USER SHOULD CAN USE THIS
-extern double __MAPPED_DY_DX__[2][N-1];         // mapped derivative values             - USER SHOULD CAN USE THIS
-extern double __INTEGRAL_XY__[2][N-1];          // integral values                      - USER SHOULD CAN USE THIS
-extern double __MAPPED_INTEGRAL_XY__[2][N-1];   // mapped integral values               - USER SHOULD CAN USE THIS
-extern double __AREA__;                         // __AREA__ under the curve             - USER SHOULD CAN USE THIS
-extern double __BISECTION__[2][MAX_ZEROS*2];    // bisection points                     - USER SHOULD CAN USE THIS
+extern float __X_MIN__;                        // minimum x value                      - USER SHOULD CHANGE THIS
+extern float __X_MAX__;                        // maximum x value                      - USER SHOULD CHANGE THIS
+extern float __Y_MIN__;                        // minimum y value                      - USER SHOULD USE THIS
+extern float __Y_MAX__;                        // maximum y value                      - USER SHOULD USE THIS
+extern float __Y_MIN_DY_DX__;                  // minimum y value of dy/dx             - USER SHOULD USE THIS
+extern float __Y_MAX_DY_DX__;                  // maximum y value of dy/dx             - USER SHOULD USE THIS
+extern float __Y_MIN_INTEGRAL__;               // minimum y value of integral of y     - USER SHOULD USE THIS
+extern float __Y_MAX_INTEGRAL__;               // maximum y value of integral of y     - USER SHOULD USE THIS
+extern float __MP_X1__;                        // x1 value for mapping                 - USER SHOULD CHANGE THIS
+extern float __MP_Y1__;                        // y1 value for mapping                 - USER SHOULD CHANGE THIS
+extern float __MP_X2__;                        // x2 value for mapping                 - USER SHOULD CHANGE THIS
+extern float __MP_Y2__;                        // y2 value for mapping                 - USER SHOULD CHANGE THIS
+extern float __ZOOMING__;                      // zooming factor                       - USER SHOULD NOT CHANGE THIS
+extern float __SHIFTING__;                     // shifting factor                      - USER SHOULD NOT CHANGE THiS
+extern float __XY__[2][N];                     // x values                             - USER SHOULD CAN USE THIS
+extern float __MAPPED_XY__[2][N];              // x values mapped to screen coordinates- USER SHOULD USE THIS
+extern float __DY_DX__[2][N-1];                // derivative values                    - USER SHOULD CAN USE THIS
+extern float __MAPPED_DY_DX__[2][N-1];         // mapped derivative values             - USER SHOULD CAN USE THIS
+extern float __INTEGRAL_XY__[2][N-1];          // integral values                      - USER SHOULD CAN USE THIS
+extern float __MAPPED_INTEGRAL_XY__[2][N-1];   // mapped integral values               - USER SHOULD CAN USE THIS
+extern float __AREA__;                         // __AREA__ under the curve             - USER SHOULD CAN USE THIS
+extern float __BISECTION__[2][MAX_ZEROS*2];    // bisection points                     - USER SHOULD CAN USE THIS
 extern Stack __ZEROS__;                         // zeros of the function                - USER SHOULD CAN USE THIS
 extern Stack __MAPPED_ZEROS__;                  // mapped zeros of the function         - USER SHOULD CAN USE THIS
 
@@ -60,19 +60,19 @@ void initialize();                                // Initialize global variables
 
 // EVALUATION OF MATHEMATICAL EXPRESSIONS
 void init(Stack* s);                                // Initialize stack
-void push(Stack* s, double val);                    // Push item onto stack
-double pop(Stack* s);                               // Pop item from stack
-double peek(Stack* s);                              // Peek at the top item of the stack
+void push(Stack* s, float val);                    // Push item onto stack
+float pop(Stack* s);                               // Pop item from stack
+float peek(Stack* s);                              // Peek at the top item of the stack
 int isempty(Stack* s);                              // Check if stack is empty
 void print_stack(Stack* s);                         // Print stack
 int precedence(char op);                            // Operator precedence
-double applyOp(double a, double b, char op);        // Apply operator to operands
-double eval(const char* expr);                      // Evaluate expression
+float applyOp(float a, float b, char op);        // Apply operator to operands
+float eval(const char* expr);                      // Evaluate expression
 
 // GENERATION OF X AND Y VALUES
 void bracket_adder();               // Envolopes sin, cos, tan, log, and abs in brackets
 void x_vals();                      // Generate x values
-void val_replacer(double val);      // Replace variable with value
+void val_replacer(float val);      // Replace variable with value
 void y_vals();                      // Generate y values corresponding to x values
 void xy_vals();                     // Generate x and y values
 
@@ -81,12 +81,12 @@ void derivative();                   // Calculate the derivative of a function
 void integral();                     // Calculate the integral of a function - Trapezoidal Rule - Integration from __X_MIN__ to x
 void area_under_curve();             // Calculate the __AREA__ under the curve of a function
 void bisection_points();             // Calculate the bisection points of a function
-double bisection_method(double point1[1][2], double point2[1][2]); // Calculate the root using the bisection method with given points  
+float bisection_method(float point1[1][2], float point2[1][2]); // Calculate the root using the bisection method with given points  
 void zeros_of_function();                // Calculate the zeros of a function
 
 // MAPPING OF X AND Y VALUES TO SCREEN COORDINATES
-double max(double* arr, int n);     // Find maximum value in an array
-double min(double* arr, int n);     // Find minimum value in an array
+float max(float* arr, int n);     // Find maximum value in an array
+float min(float* arr, int n);     // Find minimum value in an array
 void map_xy();                      // Map x and y values to screen coordinates
 void map_dx_dy();                   // Map dy/dx values to screen coordinates
 void map_integral();                // Map integral y values to screen coordinates
